[ Change ] duration in startGame Method from iMainView [ Add ] Counter layout which contains : - Progressbar - Text view [ Add ] Progressbar BackGround and Progress Drawable [ Bind ] Counter Views

+++ FallingWords/app/src/main/java/xom/xyz/fallingwords/design/model/GameModel.java
@@ -9,5 +9,5 @@ public interface GameModel {
     String getRightAnswer();
     String getProposedAnswer();
     int getScore();
-    Double getTime();
+    Long getTime();
 }
